+++ packages/isaacscript-common/src/functions/map.ts
@@ -1,3 +1,4 @@
+import { DefaultMap } from "../classes/DefaultMap";
 import { sumArray } from "./array";
 import { getPartialMatch } from "./string";
 
@@ -9,6 +10,34 @@ export function copyMap<K, V>(oldMap: Map<K, V>): Map<K, V> {
   }
 
   return newMap;
+}
+
+/**
+ * Helper function to get the value from a `DefaultMap` that corresponds to an entity, assuming that
+ * the map uses `PtrHash` as an index.
+ */
+export function defaultMapGetHash<V, A extends unknown[]>(
+  map: DefaultMap<PtrHash, V, A>,
+  entity: Entity,
+  ...extraArgs: A
+): V {
+  const ptrHash = GetPtrHash(entity);
+  return map.getAndSetDefault(ptrHash, ...extraArgs);
+}
+
+/**
+ * Helper function to set a value for a `DefaultMap` that corresponds to an entity, assuming that
+ * the map uses `PtrHash` as an index.
+ *
+ * Since `Map` and `DefaultMap` set values in the same way, this function is simply an alias for the
+ * `mapSetHash` helper function.
+ */
+export function defaultMapSetHash<V>(
+  map: Map<PtrHash, V>,
+  entity: Entity,
+  value: V,
+): void {
+  mapSetHash(map, entity, value);
 }
 
 /**
@@ -49,6 +78,19 @@ export function getMapPartialMatch<T>(
   }
 
   return [matchingKey, value];
+}
+
+/**
+ * Helper function to set a value for a `DefaultMap` that corresponds to an entity, assuming that
+ * the map uses `PtrHash` as an index.
+ */
+export function mapSetHash<V>(
+  map: Map<PtrHash, V>,
+  entity: Entity,
+  value: V,
+): void {
+  const hash = GetPtrHash(entity);
+  map.set(hash, value);
 }
 
 /** Helper function to sum every value in a map together. */